requirements.txt
amqp==2.6.0
apns2==0.7.1
appdirs==1.4.3
asgiref==3.2.3
authy==2.2.6
Babel==2.9.0
base58==2.0.1
billiard==3.6.3.0
bitcoin==1.1.39
blockchain==1.4.4
blockcypher==1.0.80
CacheControl==0.12.6
celery==4.4.4
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
cryptography==2.9.2
defusedxml==0.6.0
diff-match-patch==20181111
distlib==0.3.0
distro==1.4.0
Django==3.0
django-admin-rangefilter==0.5.2
django-axes==5.13.1
django-cors-headers==3.2.0
django-formtools==2.2
django-import-export==2.0
django-ipware==3.0.2
django-modeladmin-reorder==0.3.1
django-nested-admin==3.3.3
django-otp==1.0.2
django-phonenumber-field==3.0.1
django-push-notifications==2.0.0
django-session-security==2.6.6
django-suit==2.0a1
django-two-factor-auth==1.13
djangorestframework==3.10.3
djangorestframework-simplejwt==4.4.0
ecdsa==0.15
ed25519==1.5
enum-compat==0.0.3
et-xmlfile==1.0.1
fcm-django==0.3.7
future==0.18.2
h2==2.6.2
hpack==3.0.0
html5lib==1.0.1
http-ece==1.1.0
hyper==0.7.0
hyperframe==3.2.0
idna==2.9
ipaddr==2.2.0
jdcal==1.4.1
kombu==4.6.10
lockfile==0.12.2
MarkupPy==1.14
msgpack==0.6.2
NamedAtomicLock==1.1.3
odfpy==1.4.0
openpyxl==3.0.2
packaging==20.3
pep517==0.8.2
phonenumbers==8.12.20
progress==1.5
py-vapid==1.7.0
pycparser==2.20
pyfcm==1.4.9
PyJWT==1.7.1
pyparsing==2.4.6
python-dateutil==2.8.1
python-magic==0.4.22
python-monkey-business==1.0.0
pytoml==0.1.21
pytz==2019.3
pywebpush==1.11.0
PyYAML==5.2
qrcode==6.1
requests==2.23.0
retrying==1.3.3
six==1.14.0
sqlparse==0.3.0
tablib==0.14.0
twilio==6.52.0
urllib3==1.25.8
vine==1.3.0
webencodings==0.5.1
xlrd==1.2.0
xlwt==1.3.0
xmltodict==0.12.0
django-environ
